Separate proc into submodules
