+++ modules/image/src/img_iface.cpp
@@ -145,7 +145,7 @@ static	void	img_iface_save_file		(void);
 
 
 /******************************************************************************
- ******* main *****************************************************************
+ ******* global functions *****************************************************
  ******************************************************************************/
 void	img_iface_init		(void)
 {
