+++ modules/menu/tmp/Makefile
@@ -59,7 +59,7 @@ MENUCLUI_INC_DIRS	= -I $(INC_DIR) \
 MENUTUI_INC_LIBALX	= alx_ncur.h
 MENUTUI_INC_ABOUT	= about.h
 MENUTUI_INC_CTRL	= start.h
-MENUTUI_INC_PROC	= proc.h
+MENUTUI_INC_PROC	= proc_iface.h
 MENUTUI_INC_SAVE	= save.h
 MENUTUI_INC_USR		= user_iface.h
 MENUTUI_INC		= menu_tui.h
