+++ modules/proc/inc/proc_iface.h
@@ -1,0 +1,60 @@
+/******************************************************************************
+ *	Copyright (C) 2018	Alejandro Colomar Andrés		      *
+ ******************************************************************************/
+
+
+/******************************************************************************
+ ******* include guard ********************************************************
+ ******************************************************************************/
+# ifndef		VA_PROC_IFACE_H
+	# define	VA_PROC_IFACE_H
+
+
+/******************************************************************************
+ ******* macros ***************************************************************
+ ******************************************************************************/
+
+
+/******************************************************************************
+ ******* enums ****************************************************************
+ ******************************************************************************/
+	enum	Proc_Mode {
+		PROC_MODE_FOO,
+
+		PROC_MODE = 0x4000,
+		PROC_MODE_LABEL,
+		PROC_MODE_COINS,
+		PROC_MODE_RESISTOR
+	};
+
+	enum	Proc_DBG {
+		PROC_DBG_NO,
+		PROC_DBG_STOP_ITEM,
+		PROC_DBG_DELAY_STEP,
+		PROC_DBG_STOP_STEP
+	};
+
+
+/******************************************************************************
+ ******* variables ************************************************************
+ ******************************************************************************/
+	extern	int	proc_debug;
+	extern	int	proc_mode;
+
+
+/******************************************************************************
+ ******* functions ************************************************************
+ ******************************************************************************/
+	int	proc_iface_single	(int action);
+	void	proc_iface_series	(void);
+
+
+/******************************************************************************
+ ******* include guard ********************************************************
+ ******************************************************************************/
+# endif			/* proc_iface.h */
+
+
+/******************************************************************************
+ ******* end of file **********************************************************
+ ******************************************************************************/
