+++ modules/proc/inc/proc_label.h
@@ -6,8 +6,8 @@
 /******************************************************************************
  ******* include guard ********************************************************
  ******************************************************************************/
-# ifndef		VA_PROC_H
-	# define	VA_PROC_H
+# ifndef		VA_PROC_LABEL_H
+	# define	VA_PROC_LABEL_H
 
 
 /******************************************************************************
@@ -18,22 +18,6 @@
 /******************************************************************************
  ******* enums ****************************************************************
  ******************************************************************************/
-	enum	Proc_Mode {
-		PROC_MODE_FOO,
-
-		PROC_MODE = 0x4000,
-		PROC_MODE_LABEL,
-		PROC_MODE_COINS,
-		PROC_MODE_RESISTOR
-	};
-
-	enum	Proc_DBG {
-		PROC_DBG_NO,
-		PROC_DBG_STOP_ITEM,
-		PROC_DBG_DELAY_STEP,
-		PROC_DBG_STOP_STEP
-	};
-
 	enum	Proc_Label {
 		LABEL_OK,
 		LABEL_NOK_LABEL,
@@ -43,39 +27,22 @@
 		LABEL_NOK_PRICE
 	};
 
-	enum	Proc_Coins {
-		COINS_OK,
-		COINS_NOK_COINS,
-		COINS_NOK_OVERLAP
-	};
-
-	enum	Proc_Resistor {
-		RESISTOR_OK,
-		RESISTOR_NOK_RESISTOR,
-		RESISTOR_NOK_BANDS,
-		RESISTOR_NOK_STD_VALUE,
-		RESISTOR_NOK_TOLERANCE
-	};
-
 
 /******************************************************************************
  ******* variables ************************************************************
  ******************************************************************************/
-	extern	int	proc_debug;
-	extern	int	proc_mode;
 
 
 /******************************************************************************
  ******* functions ************************************************************
  ******************************************************************************/
-	int	proc_iface_single	(int action);
-	void	proc_iface_series	(void);
+	int	proc_label		(void);
 
 
 /******************************************************************************
  ******* include guard ********************************************************
  ******************************************************************************/
-# endif			/* proc.h */
+# endif			/* proc_label.h */
 
 
 /******************************************************************************
